requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# API and Environment
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
aiohttp>=3.8.5  # For async HTTP requests
websockets>=11.0.3  # For WebSocket server
sgp4>=2.23  # For satellite position calculation

# Machine Learning
tensorflow>=2.13.0
scikit-learn>=1.3.0
filterpy>=1.4.5  # For Kalman filters

# Visualization
plotly>=5.15.0

# Testing
pytest>=7.4.0
